SoChain limits number of transaction it searches, even if they are not different.
PageTableView now allows you to click the page number to manually type in the page you want.

When AddressExplorerActivity and AddressPortfolioExplorerActivity are loaded, a "Download Data" button is what the user should first hit to start downloading data. Then the standard table buttons appear.

AddressExplorerActivity and AddressPortfolioExplorerActivity are rewritten to use "CryptoAddress" more, so that they can function even if there is no AddressData loaded yet.